rss: Enable RSS images in FIPs
Remove RSS flash creation and instead insert the images into FIPs. Move
GPT creation into build-rss.sh so it can be run after the RSS images are
inserted. Change the alignment of the GPT partitions to 8KiB, to match
the required alignment for the RSS ATU/SIC, and patch how the GPT image
is built to allows for partitions that are aligned to 8KiB boundaries.

Signed-off-by: Raef Coles <raef.coles@arm.com>
Change-Id: Ic421ad61ad0639e8832f233df46627403503f2fb