linux: Backport support for architected qarma3 pac
On the FVP without the Crypto.so plugin this has no functional
change.

With the Crypto.so plugin or with rtl then the pac
algorithm used is the architected qarma3. Without this change
linux considers that pac is not present. With this change linux
detects that the architected qarma3 algorithm is used and the
pac kernel selftests pass.

Signed-off-by: Ben Horgan <ben.horgan@arm.com>