tc2: adopt the use of gcc-arm-12.2rel1 toolchain
Update the version of the GCC toolchain used to compile TC to version
gcc-arm-12.2rel1.

Compiling some components may cause some warnings which with the use
of "-Werror", will be treated as fatal errors and stop the compilation.
Therefore, some temporary workaround patches were applied to overcome
this error and allow build of the TC Sw stack.

Signed-off-by: Sergio Alves <sergio.dasilvalves@arm.com>