Adding ML app and models to Buildroot image.
The target machine learning applications built for Buildroot added
to the ramdisk image so they are available for users to execute
out-of-the-box. A couple of sample models (.tflite files) have also
been made available.

A new file `ml.config` has been extracted out of the `common.config`
and appended with URLs and location for these models.

Signed-off-by: Kshitij Sisodia <kshitij.sisodia@arm.com>