tc2: Use genpd framework for dvfs when using android source gpu ddk
Using the genpd framework allows us to change the frequency and
voltage together using the scmi perf protocol. This is achieved by
patching the gpu mali kbase driver to introduce a call to
dev_pm_genpd_set_performance_state.

This patch does not include patching the driver when using the
prebuilt ddk binaries or patching the driver when using it from
debian.

Signed-off-by: Ben Horgan <ben.horgan@arm.com>