tc: add a condition to check for an empty folder while cloning GPU DDK
On any new machine when host keys are not added to GPU DDK gerrit
project, git clone of DDK is bound to fail. But, even though it fails
it still creates an empty folder locally. Next time, after adding the
host keys to GPU DDK gerrit and try to clone it again, we end up not
cloning it. This is becuase we have a condition where we check if
folder exists then do not clone and proceed ahead.

This change is now also checking if folder is empty or not and
based on that we clone the GPU DDK.

Signed-off-by: Rupinderjit Singh <rupinderjit.singh@arm.com>