Revising TC_GPU to be optional
This patch changes the env variable `TC_GPU` being set to
`hwr-prebuilt` by default and removes the checks for this
variable being set for all filesystems. Instead we enforce
this variable only Android builds. This is done to be able
to boot Debian without having to build the GPU DDK.

Signed-off-by: Kshitij Sisodia <kshitij.sisodia@arm.com>