tc: fix doc config to comply with recent RTD support changes
Update default config for docs so that public RTD can successfuly
render the documentation for public releases.
This update is enforced by the recent change made on public RTD
to deprecate the use of build.image config key (not used in TC docs)
and enforce the use of build.os config key instead (now mandatory for
all projects, since Oct 16, 2023).

Signed-off-by: Sergio Alves <sergio.dasilvalves@arm.com>